Added support for private APIs
Added @apiPrivate to support the ability to specify
if an API is a private API.  apidoc now also has a
--private false|true parameter on the command line
that defaults to false.  If false, it does not include
any API block that contains @apiPrivate.  If true, it
will include the API (assuming the API is also not marked
as @apiIgnore).