Fix the regex parsing for valueless parameters
The @apiPrivate parameter does not take any value, and the
@apiDeprecated parameter has an optional value. However, the regular
expression in Parser.prototype.findElements expects that every parameter
has a value. This may lead to the subsequent apiDoc parameter getting
matched into the valueless parameter, removing it from the final
generated apiDocs.

Switching the regular expression from (.+?) to (.*?) allows the
parameter's value to be optional.