Modify parser to use first match even if in higher positions.
When using Elixir with single quote heredoc syntax, eg
    @apidoc '''
    @api {get} /index
    '''
The docBlocksRegExp match is in the element 4 of the matches array.
An error was being thrown because only elements 1 & 2 were being
checked for a match.

Since a match is guaranteed, otherwise the regular expression would
have returned null and the loop would exit, it seems appropriate to
use the first non-null element of the matches array as a block. The
first element is excluded though, as this always contains the full
string being checked rather than a substring match.